refactor(mouse): move tmux gesture + arbiter under input::mouse::button::tmux
Relocate the tmux left-button gesture pipeline from src/input/tmux/mouse/
to src/input/mouse/button/tmux/, and fold the tmux webview pointer arbiter
(tmux_webview_pointer, forward_tmux_webview_mouse_moves) into the gesture
file so the state machine + effect IR stay in one subtree. Rename
MousePlugin -> MouseButtonTmuxPlugin (registered by MouseButtonInputPlugin);
widen pane_hit::tmux_pane_at_phys to pub(in crate::input); re-export
divider_at through crate::input::mouse. Behavior-preserving.
