+++ src/input/mouse/button.rs
@@ -27,7 +27,10 @@ use ozma_tty_engine::{
 use ozma_tty_renderer::TerminalCellMetricsResource;
 use std::time::Duration;
 
-/// Registers the mouse-button dispatcher and its gesture resource. Runs in
+pub(in crate::input::mouse) mod tmux;
+
+/// Registers the mouse-button dispatcher and its gesture resource, plus the
+/// tmux left-button gesture pipeline (`tmux::MouseButtonTmuxPlugin`). Runs in
 /// `InputPhase::Dispatch`, gated to frames carrying any mouse message — the
 /// focus/empty-candidate guard must still run on wheel-only frames to drain
 /// readers and reset the gesture.
@@ -35,12 +38,14 @@ pub(super) struct MouseButtonInputPlugin;
 
 impl Plugin for MouseButtonInputPlugin {
     fn build(&self, app: &mut App) {
-        app.init_resource::<OzmaMouseGesture>().add_systems(
-            Update,
-            dispatch_mouse_buttons
-                .in_set(InputPhase::Dispatch)
-                .run_if(on_any_mouse_message()),
-        );
+        app.init_resource::<OzmaMouseGesture>()
+            .add_systems(
+                Update,
+                dispatch_mouse_buttons
+                    .in_set(InputPhase::Dispatch)
+                    .run_if(on_any_mouse_message()),
+            )
+            .add_plugins(tmux::MouseButtonTmuxPlugin);
     }
 }
